First Push
So far, I added a test image just to see if drawing a background works. When it runs, it opens a jframe, and then you can click the 'new game' button to add the background to the scene. General structure has also been added to the Racer class, and is ready for additional game components to be added to it.